Made record.sh a bit more robust

+++ record.sh
@@ -3,8 +3,13 @@
 FILE=$1
 RDIR=records/$FILE
 
-mkdir $RDIR
+mkdir -p $RDIR
 
-cp $FILE.txt $RDIR/raw.txt
-./analyze.py < $FILE.txt > $RDIR/analyzed.txt
-./aggregate.py < $RDIR/analyzed.txt > $RDIR/aggregated.txt
+if [ -e $FILE.txt ]; then
+    cp $FILE.txt $RDIR/raw.txt
+    ./analyze.py < $FILE.txt > $RDIR/analyzed.txt
+    ./aggregate.py < $RDIR/analyzed.txt > $RDIR/aggregated.txt
+
+else
+    echo "Couldn't find file $FILE.txt"
+fi
